requirement.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
Pillow
PyJWT
pytz
sqlparse
